Move tests related to the `Identifiers` class to the `IdentifiersShould` class.
