+++ client/src/test/java/org/spine3/convert/StringifiersShould.java
@@ -18,23 +18,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.base;
+package org.spine3.convert;
 
 import com.google.common.reflect.TypeToken;
-import com.google.protobuf.Any;
-import com.google.protobuf.Int32Value;
-import com.google.protobuf.Int64Value;
-import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
 import org.junit.Test;
-import org.spine3.protobuf.AnyPacker;
+import org.spine3.base.CommandId;
+import org.spine3.base.Commands;
+import org.spine3.base.EventId;
+import org.spine3.base.Events;
 import org.spine3.protobuf.Timestamps2;
 import org.spine3.test.NullToleranceTest;
 import org.spine3.test.identifiers.IdWithPrimitiveFields;
-import org.spine3.test.identifiers.NestedMessageId;
-import org.spine3.test.identifiers.SeveralFieldsId;
-import org.spine3.test.identifiers.TimestampFieldId;
 
 import java.text.ParseException;
 import java.util.Random;
@@ -43,157 +39,15 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.base.Identifiers.EMPTY_ID;
-import static org.spine3.base.Identifiers.NULL_ID;
 import static org.spine3.base.Identifiers.idToString;
-import static org.spine3.base.Identifiers.toIdString;
-import static org.spine3.protobuf.Timestamps2.getCurrentTime;
-import static org.spine3.protobuf.Values.newIntValue;
-import static org.spine3.protobuf.Values.newLongValue;
-import static org.spine3.protobuf.Values.newStringValue;
+import static org.spine3.base.Identifiers.newUuid;
 import static org.spine3.test.Tests.assertHasPrivateParameterlessCtor;
 
 public class StringifiersShould {
 
-    private static final String TEST_ID = "someTestId 1234567890 !@#$%^&()[]{}-+=_";
-
     @Test
     public void have_private_constructor() {
         assertHasPrivateParameterlessCtor(Stringifiers.class);
-    }
-
-    @Test
-    public void return_NULL_ID_if_convert_null_to_string() {
-        assertEquals(NULL_ID, idToString(null));
-    }
-
-    @Test
-    public void return_EMPTY_ID_if_convert_empty_string_to_string() {
-        assertEquals(EMPTY_ID, idToString(""));
-    }
-
-    @Test
-    public void return_EMPTY_ID_if_convert_blank_string_to_string() {
-        assertEquals(EMPTY_ID, idToString(" "));
-    }
-
-    @Test
-    public void return_EMPTY_ID_if_result_of_Message_to_string_conversion_is_empty_string() {
-        assertEquals(EMPTY_ID, idToString(CommandId.getDefaultInstance()));
-    }
-
-    @Test
-    public void return_EMPTY_ID_if_result_of_Message_to_string_conversion_is_blank_string() {
-        assertEquals(EMPTY_ID, idToString(newStringValue("  ")));
-    }
-
-    @Test
-    public void return_EMPTY_ID_if_convert_empty_message_to_string() {
-        assertEquals(EMPTY_ID, idToString(StringValue.getDefaultInstance()));
-    }
-
-    @Test
-    public void return_string_id_as_is() {
-        assertEquals(TEST_ID, idToString(TEST_ID));
-    }
-
-    @Test
-    public void return_same_string_when_convert_string_wrapped_into_message() {
-
-        final StringValue id = newStringValue(TEST_ID);
-
-        final String result = idToString(id);
-
-        assertEquals(TEST_ID, result);
-    }
-
-    @Test
-    public void convert_to_string_integer_id_wrapped_into_message() {
-        final Integer value = 1024;
-        final Int32Value id = newIntValue(value);
-        final String expected = value.toString();
-
-        final String actual = idToString(id);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void convert_to_string_long_id_wrapped_into_message() {
-        final Long value = 100500L;
-        final Int64Value id = newLongValue(value);
-        final String expected = value.toString();
-
-        final String actual = idToString(id);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void convert_to_string_message_id_with_string_field() {
-        final StringValue id = newStringValue(TEST_ID);
-
-        final String result = idToString(id);
-
-        assertEquals(TEST_ID, result);
-    }
-
-    @Test
-    public void convert_to_string_message_id_with_timestamp_field() {
-        final Timestamp currentTime = getCurrentTime();
-        final TimestampFieldId id = TimestampFieldId.newBuilder()
-                                                    .setId(currentTime)
-                                                    .build();
-        final String expected = toIdString(currentTime);
-
-        final String actual = idToString(id);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void convert_to_string_message_id_with_message_field() {
-        final StringValue value = newStringValue(TEST_ID);
-        final NestedMessageId idToConvert = NestedMessageId.newBuilder()
-                                                           .setId(value)
-                                                           .build();
-
-        final String result = idToString(idToConvert);
-
-        assertEquals(TEST_ID, result);
-    }
-
-    @Test
-    public void have_default_to_string_conversion_of_message_id_with_several_fields() {
-        final String nestedString = "nested_string";
-        final String outerString = "outer_string";
-        final Integer number = 256;
-
-        final SeveralFieldsId idToConvert = SeveralFieldsId.newBuilder()
-                                                           .setString(outerString)
-                                                           .setInt(number)
-                                                           .setMessage(newStringValue(nestedString))
-                                                           .build();
-
-        final String expected =
-                "string=\"" + outerString + '\"' +
-                " int=" + number +
-                " message { value=\"" + nestedString + "\" }";
-
-        final String actual = idToString(idToConvert);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void convert_to_string_message_id_wrapped_in_Any() {
-        final StringValue messageToWrap = newStringValue(TEST_ID);
-        final Any any = AnyPacker.pack(messageToWrap);
-
-        final String result = idToString(any);
-
-        assertEquals(TEST_ID, result);
     }
 
     private static final Stringifier<IdWithPrimitiveFields> ID_TO_STRING_CONVERTER =
@@ -217,12 +71,13 @@ public class StringifiersShould {
                            .register(TypeToken.of(IdWithPrimitiveFields.class),
                                      ID_TO_STRING_CONVERTER);
 
+        final String testId = "testId 123456";
         final IdWithPrimitiveFields id = IdWithPrimitiveFields.newBuilder()
-                                                              .setName(TEST_ID)
+                                                              .setName(testId)
                                                               .build();
         final String result = idToString(id);
 
-        assertEquals(TEST_ID, result);
+        assertEquals(testId, result);
     }
 
     @SuppressWarnings("OptionalGetWithoutIsPresent")
